requirements.txt
moviepy==1.0.3
pillow==10.4.0
openai==1.61.1
python-dotenv==1.0.1